task projet view

+++ app/src/main/java/com/mlk/taskmanager/data/repository/ProjectRepository.kt
@@ -6,15 +6,15 @@ import kotlinx.coroutines.flow.Flow
 interface ProjectRepository {
     fun getAllProjects(): Flow<List<Project>>
     
-    suspend fun getProjectById(id: Int): Project?
+    suspend fun getProjectById(id: Long): Project?
     
     suspend fun insertProject(project: Project): Long
     
     suspend fun updateProject(project: Project)
     
-    suspend fun deleteProject(id: Int)
+    suspend fun deleteProject(id: Long)
     
-    suspend fun incrementTaskCount(projectId: Int)
+    suspend fun incrementTaskCount(projectId: Long)
     
-    suspend fun decrementTaskCount(projectId: Int)
+    suspend fun decrementTaskCount(projectId: Long)
 }
